moved all new content to project-1 subdirectory
